+++ gitpush.sh
@@ -1,3 +1,3 @@
 git add .
-git commit -m "jul 17, changed changes.md"
+git commit -m "jul 19, eslint updated configuration, added .prettierrc and .eslintignore"
 git push -u origin main