+++ src/App.tsx
@@ -1,9 +1,7 @@
-import React from "react";
-import "./aoo.css";
-import { AdderCom } from "@Components";
+import React from 'react';
+import './aoo.css';
+import { AdderCom } from '@Components';
 
-const App = () => {
-  return <AdderCom />;
-};
+const App = () => <AdderCom />;
 
 export default App;