+++ src/Store/index.ts
@@ -1,3 +1,3 @@
-export * from "./store";
-export * from "./DispatchHooks";
-export * from "./Features/counter/counterSlice";
+export * from './store';
+export * from './DispatchHooks';
+export * from './Features/counter/counterSlice';