+++ src/index.tsx
@@ -1,19 +1,18 @@
-import { Provider } from "react-redux";
-import { store } from "./Store/store";
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
-import ErrorBoundary from "./ErrorLogger/ErrorLogger";
-
-import { QueryClient, QueryClientProvider } from "react-query";
-import { ReactQueryDevtools } from "react-query/devtools";
+import { Provider } from 'react-redux';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { ReactQueryDevtools } from 'react-query/devtools';
+import { store } from './Store/store';
+import App from './App';
+import ErrorBoundary from './ErrorLogger/ErrorLogger';
 
 // react-query client
 const queryClient = new QueryClient();
 
-const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
-);
+const rootElement = document.getElementById('root') as HTMLElement;
+
+const root = ReactDOM.createRoot(rootElement);
 root.render(
   <QueryClientProvider client={queryClient}>
     <Provider store={store}>